refactor: fix the architecture
